--- example/web/apitest.html
@@ -1,211 +1,0 @@
-<!DOCTYPE html>
-
-<!-- Copyright (c) 2014, the Dart project authors.  Please see the AUTHORS file
-     for details. All rights reserved. Use of this source code is governed by a
-     BSD-style license that can be found in the LICENSE file. -->
-
-<html>
-  <head>
-  	<meta charset="utf-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-  	<meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>API Test</title>
-    <link rel="icon" href="favicon.ico">
-    <script defer src="apitest.dart.js"></script>
-    <link href="apitest.css" rel="stylesheet">
-    <link href="packages/codemirror/codemirror.css" rel="stylesheet">
-  </head>
-
-  <body>
-    <script src="packages/codemirror/codemirror.js"></script>
-
-    <h2>Dartpad Server API Tester</h2>
-    <!--<input type="text" value="https://dart-services.appspot.com/">-->
-    <input type="text" value="http://localhost:8082/api">
-
-    <section id="analyzeSection-v1">
-      <h3>api/dartservices/v1/analyze</h3>
-      <div class="row">
-        <button class="post">POST</button>
-      </div>
-      <div class="container">
-        <div class="editor"></div>
-        <div class="output"></div>
-      </div>
-    </section>
-
-    <section id="analyzeSection-v2">
-      <h3>api/dartservices/v2/analyze</h3>
-      <div class="row">
-        <button class="post">POST</button>
-      </div>
-      <div class="container">
-        <div class="editor"></div>
-        <div class="output"></div>
-      </div>
-    </section>
-
-    <section id="assistsSection-v1">
-      <h3>api/dartservices/v1/assists</h3>
-      <div class="row">
-        <button class="post">POST</button>
-        <span class="offset"></span>
-      </div>
-      <div class="container">
-        <div class="editor"></div>
-        <div class="output"></div>
-      </div>
-    </section>
-
-    <section id="assistsSection-v2">
-      <h3>api/dartservices/v2/assists</h3>
-      <div class="row">
-        <button class="post">POST</button>
-        <span class="offset"></span>
-      </div>
-      <div class="container">
-        <div class="editor"></div>
-        <div class="output"></div>
-      </div>
-    </section>
-
-    <section id="compileSection-v1">
-      <h3>api/dartservices/v1/compile</h3>
-      <div class="row">
-        <button class="post">POST</button>
-      </div>
-      <div class="container">
-        <div class="editor"></div>
-        <div class="output"></div>
-      </div>
-    </section>
-
-    <section id="compileSection-v2">
-      <h3>api/dartservices/v2/compile</h3>
-      <div class="row">
-        <button class="post">POST</button>
-      </div>
-      <div class="container">
-        <div class="editor"></div>
-        <div class="output"></div>
-      </div>
-    </section>
-
-    <section id="compileDDCSection-v1">
-      <h3>api/dartservices/v1/compileDDC</h3>
-      <div class="row">
-        <button class="post">POST</button>
-      </div>
-      <div class="container">
-        <div class="editor"></div>
-        <div class="output"></div>
-      </div>
-    </section>
-
-    <section id="compileDDCSection-v2">
-      <h3>api/dartservices/v2/compileDDC</h3>
-      <div class="row">
-        <button class="post">POST</button>
-      </div>
-      <div class="container">
-        <div class="editor"></div>
-        <div class="output"></div>
-      </div>
-    </section>
-
-    <section id="completeSection-v1">
-      <h3>api/dartservices/v1/complete</h3>
-      <div class="row">
-        <button class="post">POST</button>
-        <span class="offset"></span>
-      </div>
-      <div class="container">
-        <div class="editor"></div>
-        <div class="output"></div>
-      </div>
-    </section>
-
-    <section id="completeSection-v2">
-      <h3>api/dartservices/v2/complete</h3>
-      <div class="row">
-        <button class="post">POST</button>
-        <span class="offset"></span>
-      </div>
-      <div class="container">
-        <div class="editor"></div>
-        <div class="output"></div>
-      </div>
-    </section>
-
-    <section id="documentSection-v1">
-      <h3>api/dartservices/v1/document</h3>
-      <div class="row">
-        <button class="post">POST</button>
-        <span class="offset"></span>
-      </div>
-      <div class="container">
-        <div class="editor"></div>
-        <div class="output"></div>
-      </div>
-    </section>
-
-    <section id="documentSection-v2">
-      <h3>api/dartservices/v2/document</h3>
-      <div class="row">
-        <button class="post">POST</button>
-        <span class="offset"></span>
-      </div>
-      <div class="container">
-        <div class="editor"></div>
-        <div class="output"></div>
-      </div>
-    </section>
-
-    <section id="fixesSection-v1">
-      <h3>api/dartservices/v1/fixes</h3>
-      <div class="row">
-        <button class="post">POST</button>
-        <span class="offset"></span>
-      </div>
-      <div class="container">
-        <div class="editor"></div>
-        <div class="output"></div>
-      </div>
-    </section>
-
-    <section id="fixesSection-v2">
-      <h3>api/dartservices/v2/fixes</h3>
-      <div class="row">
-        <button class="post">POST</button>
-        <span class="offset"></span>
-      </div>
-      <div class="container">
-        <div class="editor"></div>
-        <div class="output"></div>
-      </div>
-    </section>
-
-    <section id="versionSection-v1">
-      <h3>api/dartservices/v1/version</h3>
-      <div class="row">
-        <button class="post">GET</button>
-      </div>
-      <div class="container">
-        <div class="editor"></div>
-        <div class="output"></div>
-      </div>
-    </section>
-    
-    <section id="versionSection-v2">
-      <h3>api/dartservices/v2/version</h3>
-      <div class="row">
-        <button class="post">GET</button>
-      </div>
-      <div class="container">
-        <div class="editor"></div>
-        <div class="output"></div>
-      </div>
-    </section>
-    
-  </body>
-</html>
