+++ example/web/styles.css
@@ -1,14 +1,55 @@
-@import url(https://fonts.googleapis.com/css?family=Roboto);
+/* Copyright (c) 2014, the Dart project authors.  Please see the AUTHORS file */
+/* for details. All rights reserved. Use of this source code is governed by a */
+/* BSD-style license that can be found in the LICENSE file. */
 
-html, body {
-  width: 100%;
-  height: 100%;
-  margin: 0;
-  padding: 0;
-  font-family: 'Roboto', sans-serif;
+body {
+  display: flex;
+  flex-direction: column;
+  margin: 15px;
+  color: #555;
 }
 
-#output {
-  padding: 20px;
-  text-align: center;
+section {
+  margin-top: 20px;
+  padding-top: 6px;
+  border-top: 1px solid #eee;
+  border-width: 2px;
+  padding-left: 16px;
+}
+
+h3 {
+  margin-top: 0;
+  margin-bottom: 8px;
+  margin-left: -16px;
+}
+
+#controls {
+  margin-bottom: 1em;
+}
+
+div.container {
+  display: flex;
+  flex-direction: row;
+}
+
+div.row {
+  margin: 0 0 6px 0;
+}
+
+.editor {
+  flex: 20;
+  display: inline-block;
+  box-shadow: 1px 1px 2px 0px rgba(50, 50, 50, 0.75);
+  margin-right: 16px;
+  height: 300px;
+}
+
+.output {
+  flex: 15;
+  display: inline-block;
+  box-shadow: 1px 1px 2px 0px rgba(50, 50, 50, 0.75);
+  height: 300px;
+  white-space: pre-wrap;
+  overflow-y: scroll;
+  font-family: monospace;
 }
